move async header and fix build
